Split single DialogInjector
